feat: metrics
- [Participation rate](https://github.com/limechain/hashport-validator/issues/350)
- [Fee Account](https://github.com/limechain/hashport-validator/issues/351)
- [Validators number & health](https://github.com/limechain/hashport-validator/issues/355)
- [Minted tokens](https://github.com/limechain/hashport-validator/issues/354)
- [Locked Tokens](https://github.com/limechain/hashport-validator/issues/353)
- [Operator account balance metric](https://github.com/limechain/hashport-validator/issues/360)
- [Success rate](https://github.com/limechain/hashport-validator/issues/356)
- [Balance between networks](https://github.com/limechain/hashport-validator/issues/357)
- [Implementation of Alerts](https://github.com/limechain/hashport-validator/issues/363)
- [Update licenses in hashport-validator files](https://github.com/limechain/hashport-validator/issues/386)
- [Change chainIds to follow conventions](https://github.com/limechain/hashport-validator/issues/395)

+++ constants/hedera.go
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 LimeChain Ltd.
+ * Copyright 2022 LimeChain Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,7 +18,7 @@ package constants
 
 const (
 	Hbar            = "HBAR"
-	HederaNetworkId = 0
+	HederaNetworkId = uint64(0)
 )
 
 // Handler topics
